reader/writer test's

+++ Hotel/src/hotel/Client.java
@@ -9,7 +9,7 @@ abstract class Client {
     public String Name;
     public String Surname;
     public int age;
-    public int PESEL;
+    public long PESEL;
     public double discount;
 
     public String getName() {
@@ -24,7 +24,7 @@ abstract class Client {
         return age;
     }
 
-    public int getPESEL() {
+    public long getPESEL() {
         return PESEL;
     }
 
@@ -34,7 +34,7 @@ abstract class Client {
 }
 
 class Person extends Client{
-    public Person(String Name, String Surname, int age, int PESEL){
+    public Person(String Name, String Surname, int age, long PESEL){
         super();
         this.Name = Name;
         this.Surname = Surname;
@@ -45,7 +45,7 @@ class Person extends Client{
 }
 
 class LoyalClient extends Client{
-    public LoyalClient(String Name, String Surname, int age, int PESEL){
+    public LoyalClient(String Name, String Surname, int age, long PESEL){
         super();
         this.Name = Name;
         this.Surname = Surname;
@@ -56,7 +56,7 @@ class LoyalClient extends Client{
 }
 
 class Learner extends Client{
-    public Learner(String Name, String Surname, int age, int PESEL){
+    public Learner(String Name, String Surname, int age, long PESEL){
         super();
         this.Name = Name;
         this.Surname = Surname;
@@ -69,7 +69,7 @@ class Learner extends Client{
 }
 
 class Student extends Client{
-    public Student(String Name, String Surname, int age, int PESEL){
+    public Student(String Name, String Surname, int age, long PESEL){
         super();
         this.Name = Name;
         this.Surname = Surname;
@@ -82,7 +82,7 @@ class Student extends Client{
 }
 
 class Pensioner extends Client{
-    public Pensioner(String Name, String Surname, int age, int PESEL){
+    public Pensioner(String Name, String Surname, int age, long PESEL){
         super();
         this.Name = Name;
         this.Surname = Surname;
@@ -93,7 +93,7 @@ class Pensioner extends Client{
 }
 
 class Invalid extends Client{
-    public Invalid(String Name, String Surname, int age, int PESEL){
+    public Invalid(String Name, String Surname, int age, long PESEL){
         super();
         this.Name = Name;
         this.Surname = Surname;
@@ -104,7 +104,7 @@ class Invalid extends Client{
 }
 
 class CompanyAgent extends Client {
-    public CompanyAgent(String Name, String Surname, int age, int PESEL) {
+    public CompanyAgent(String Name, String Surname, int age, long PESEL) {
         super();
         this.Name = Name;
         this.Surname = Surname;
